+++ dbConnect.php
@@ -1,11 +1,5 @@
 <?php
 require_once("dbDetails.php");
-
-$conn=new mysqli($serverName,$dbUser,$dbPassword,$dbName);
-
-if($conn->connect_error)
-{
-    die("Connection failed : ".$conn->connect_error);
-}
-
+require_once("functions.php");
+$conn=connectionDB($serverName,$dbUser,$dbPassword,$dbName);
 ?>
